fix(sun): Use Melba's functions for determining the sun position.
The sun position is now almost the same as in Melba (dot: 0.99999), there seems to be a ~20 seconds difference in my time and server time that could explain the difference.
This does not actually fix the issue - the sun pos seems the same now in terms of coordinates, but still sun does not light the same parts of the planet, though it's closer now.
It could be the planet is rotated, the texture is different or the projection is different.